Move the location of the review inifite list services
