+++ packages/review/src/components/infinite-list/popular-reviews-infinite.tsx
@@ -1,7 +1,7 @@
-import { useInfinitePopularReviews } from '../../services'
 import type { SortingType } from '../sorting-context'
 
 import { InfiniteList } from './infinite-list'
+import { useInfinitePopularReviews } from './services'
 import type { InfiniteReviewProps } from './types'
 
 export function PopularReviewsInfinite({
